+++ config/routes.rb
@@ -16,6 +16,7 @@ Rails.application.routes.draw do
   patch "password-reset/:token" => "password_resets#update", as: :password_reset_token
 
   get "dashboard" => "dashboard#show", as: :dashboard
+  resources :waterfalls, only: %i[index show], param: :slugged_public_id
 
-  root "dashboard#show"
+  root "waterfalls#index"
 end